Added unit testing via pytest module. Tests for TASK, SIGNAL, ID, PROCEDURES, RESULT methods. Coverage ~38%

+++ code/procedures.py
@@ -1,12 +1,15 @@
+# == Procedures for tests =============
+
 def add(a, b):
-	return a + b
+	return float(a + b)
 
 def sub(a, b):
-	return a - b
+	return float(a - b)
 
 def multiple(a, b):
-	return a * b
+	return float(a * b)
 
 def divide(a, b):
-	return a / b
+	return float(a / b)
 
+# =====================================
